Again flipflop on how we're handling this by reverting 2d96d8743940a567e99d584f9dab9883b34e843d (which itself was a revert!).
Now the translation team are using Google Translate judiciously and evidently it can sometimes add in quotes and other things which break javascript if this plugin injects in the wrong thing.

It's really tricky for translators because sometimes, a non-escaped backslash or quotes will not do anything untoward. How the translation nugget is used in the source code is the determining factor of whether a non-escaped character will cause problems, and obviously the translator can't know when that might apply.

So yes, while they should be following the source where appropriate, they should have the freedom to use whatever characters they like (within reason) without fear of breaking the entire website. Some languages even use completely different characters for quotes.